+++ genvalidity-path/test/Data/GenValidity/PathSpec.hs
@@ -14,6 +14,10 @@ import Data.GenValidity.Path ()
 spec :: Spec
 spec = do
     genValiditySpec @(Path Abs File)
+    shrinkValiditySpec @(Path Abs File)
     genValiditySpec @(Path Abs Dir)
+    shrinkValiditySpec @(Path Abs Dir)
     genValiditySpec @(Path Rel File)
+    shrinkValiditySpec @(Path Rel File)
     genValiditySpec @(Path Rel Dir)
+    shrinkValiditySpec @(Path Rel Dir)