+++ genvalidity-scientific/test/Data/GenValidity/ScientificSpec.hs
@@ -1,17 +1,17 @@
+{-# LANGUAGE TypeApplications #-}
+
 module Data.GenValidity.ScientificSpec
     ( spec
     ) where
 
 import Test.Hspec
-import Test.QuickCheck
+import Test.Validity
 
-import Data.GenValidity
 import Data.GenValidity.Scientific ()
 
 import Data.Scientific
 
 spec :: Spec
-spec =
-    describe "genValid" $
-    it "generates valid Scientific's" $
-    forAll (genValid :: Gen Scientific) isValid
+spec = do
+    genValidSpec @Scientific
+    shrinkValidSpec @Scientific